finalização dos exercicios

+++ aulas 1/hora de codar2.c
@@ -14,9 +14,9 @@ int main() {
     printf("digite a sua terceira nota:  \n");
     scanf("%d", &nota3);
 
-    media = (nota1 + nota2 + nota3) / 3;
+    media = (float)(nota1 + nota2 + nota3) / 3;
 
-    printf("A media é : %f\n", media);
+    printf("A media é : %.1f\n", media);
